Fix lower priority bugs in 9 components
- light: early return in Fireworks effect when strip has fewer than 2
  LEDs to prevent OOB access on it[0]/it[1]/it[last-1]
- rtttl: bounds check position_ in separator-skipping loop to prevent
  reading past string end
- esp32_camera: null check esp_camera_sensor_get() return in
  dump_config() and update_camera_parameters()
- sx127x: null check dio0_pin_ before digital_read() in transmit loop
- mopeka_std_check: change val[4] to val[3] to match 23-byte
  manufacturer data (struct was 24 bytes, causing 1-byte overread)
- bl0906: return early on read_array failure instead of publishing 0.0
- bmi160: fix gyro wake-up timeout from 10ms to 100ms (datasheet
  requires 51-81ms minimum)
- mics_4514: use signed arithmetic for power-ox/power-red subtractions
  to prevent uint16_t underflow corrupting calibration
- ld2450: widen convert_seconds_to_ms return and timeout_ to uint32_t
  to prevent truncation for timeouts > 65 seconds

+++ esphome/components/mics_4514/mics_4514.cpp
@@ -54,8 +54,8 @@ void MICS4514Component::update() {
 
   if (this->initial_) {
     this->initial_ = false;
-    this->ox_calibration_ = (float) (power - ox);
-    this->red_calibration_ = (float) (power - red);
+    this->ox_calibration_ = (float) ((int32_t) power - (int32_t) ox);
+    this->red_calibration_ = (float) ((int32_t) power - (int32_t) red);
     return;
   }
 
@@ -65,8 +65,8 @@ void MICS4514Component::update() {
     this->initial_ = true;
     return;
   }
-  float red_f = (float) (power - red) / this->red_calibration_;
-  float ox_f = (float) (power - ox) / this->ox_calibration_;
+  float red_f = (float) ((int32_t) power - (int32_t) red) / this->red_calibration_;
+  float ox_f = (float) ((int32_t) power - (int32_t) ox) / this->ox_calibration_;
 
   if (this->carbon_monoxide_sensor_ != nullptr) {
     float co = 0.0f;
